[Refactor] #963 Moved *_singing_*() from player-status.cpp/h to spells-song.cpp/h
