+++ src/dungeon/dungeon-processor.cpp
@@ -30,10 +30,10 @@
 #include "monster/monster-status.h"
 #include "monster/monster-util.h"
 #include "pet/pet-util.h"
-#include "player/player-status.h"
 #include "player/special-defense-types.h"
 #include "realm/realm-song-numbers.h"
 #include "realm/realm-song.h"
+#include "spell-realm/spells-song.h"
 #include "system/floor-type-definition.h"
 #include "system/monster-race-definition.h"
 #include "system/player-type-definition.h"
