+++ src/spell-kind/spells-detection.cpp
@@ -20,9 +20,9 @@
 #include "object-hook/hook-enchant.h"
 #include "object/object-mark-types.h"
 #include "object/tval-types.h"
-#include "player/player-status.h"
 #include "realm/realm-song-numbers.h"
 #include "realm/realm-song.h"
+#include "spell-realm/spells-song.h"
 #include "system/floor-type-definition.h"
 #include "system/monster-race-definition.h"
 #include "system/object-type-definition.h"
