+++ src/status/body-improvement.cpp
@@ -9,6 +9,7 @@
 #include "player-info/avatar.h"
 #include "player/player-status.h"
 #include "realm/realm-song-numbers.h"
+#include "spell-realm/spells-song.h"
 #include "system/player-type-definition.h"
 #include "view/display-messages.h"
