+++ src/status/element-resistance.cpp
@@ -4,9 +4,9 @@
 #include "core/stuff-handler.h"
 #include "game-option/disturbance-options.h"
 #include "player/player-race.h"
-#include "player/player-status.h"
 #include "player/special-defense-types.h"
 #include "realm/realm-song-numbers.h"
+#include "spell-realm/spells-song.h"
 #include "system/player-type-definition.h"
 #include "view/display-messages.h"
